requirements.txt
pandas
numpy
requests
streamlit
plotly
scikit-learn
python-dotenv
lxml
matplotlib
seaborn